v8: extend scope of extension to handle platform
Redesign: Previous design used node:vm module and lock/unlock
functions to handle the protected context access.

Apparently, if an exception occurs within running script, node:vm
performs checkout of the microtask queue. This causes execution of code
within the "finally" that can rethrow the exception, preventing
"finally" block to be executed properly. As a result, the lock stack is
compromised.

The new design examines the option of create pure context without the
node environment and represent that context with its global (or any
other object created in that context). Execution of any user code will
be implemented in the v8 API, instead of the node:vm. This might also
allow static resolving of v8::Module dependencies, which might be
required for the DOM loading of remote scripts.