wip: fix context wrapper retirement issue
Context lifetime is complex:
The lifetime can be semi-controlled by dispose() method.
If the object goes out of scope it *can* be destroyed.
The functions created on the context can be stored and called later,
which can keep the context global and the context alive.
The dispose() might not destroy the context, if functions are stored,
but it stops processing microtasks.

TODO: Remove the reference from Global to container object on disposal.