wip: object observer and microtask
- Observing and deleting all wrappers on exit:
Some objects (currently only MicrotaskQueue) modifies
the isolate state when created. If that state is not restored,
by the time the isolate is destroyed, crash occurs. This can happen if
worker thread exits or on the exit of main node.
AtExit handler installed on the node::Environment should guarantee
clearing all objects before the isolate is destroyed.