wip: exposing template API
Template API is different, it seems a lot of options cannot be changed
once a template is instantiated, so for JS wrappers, all options are
given to the constructor and remain unchanged.
Unlike V8 API, the option given to the constructor can be retrieved
later, although they cannot be retrieved from the underlying template,
so wrappers must store the options before they are given to the
template.