feat: context
Context: Now the current context can be retrieved. Wrappers to context
will be applied and provided to callback functions for ObjectTemplate
like named/indexed getter, setter, query, descriptor, enumerator, definer,
deleter, etc.

There will be a UserContext extension to Context that provide ability to
run a user code (including from microtasks) within a limited amount of time,
interrupting the user code and returning the control to the designated
control function if that time is exceeded.